+++ backends/arm/_passes/__init__.py
@@ -6,12 +6,17 @@
 
 from . import arm_pass_utils  # noqa
 from .arm_pass import ArmOpTargetedPass, ArmPass  # noqa  # usort: skip
-from .accumulate_index_put_pass import AccumulateIndexPutPass  # noqa
-from .broadcast_args_pass import BroadcastArgsPass  # noqa
-from .canonicalize_gather_pass import CanonicalizeGatherPass  # noqa
 from executorch.backends.transforms.canonicalize_view_copy_permute_pass import (  # noqa
     CanonicalizeViewCopyPermutePass,
 )
+from executorch.backends.transforms.fuse_identical_input_transforms_pass import (  # noqa
+    FuseIdenticalInputTransformsPass,
+    NormalizeTransformInputPlaceholdersPass,
+)
+
+from .accumulate_index_put_pass import AccumulateIndexPutPass  # noqa
+from .broadcast_args_pass import BroadcastArgsPass  # noqa
+from .canonicalize_gather_pass import CanonicalizeGatherPass  # noqa
 from .cast_int64_pass import CastInt64BuffersToInt32Pass  # noqa
 from .cast_to_int32_pass import CastToInt32Pass  # noqa
 from .constant_folding_pass import ConstantFoldingPass  # noqa
@@ -132,10 +137,6 @@ from .fuse_constant_ops_pass import (  # noqa
 )
 from .fuse_duplicate_users_pass import FuseDuplicateUsersPass  # noqa
 from .fuse_equal_placeholders_pass import FuseEqualPlaceholdersPass  # noqa
-from .fuse_identical_input_transforms_pass import (  # noqa
-    FuseIdenticalInputTransformsPass,
-    NormalizeTransformInputPlaceholdersPass,
-)
 from .fuse_quantized_activation_pass import FuseQuantizedActivationPass  # noqa
 from .fuse_view_copy_transform_pass import FuseViewCopyTransformPass  # noqa
 from .insert_const_shapes import InsertConstShapesPass  # noqa
